alta actividad deportiva y alta de dictado de clase

+++ src/main/java/logica/ActividadDeportiva.java
@@ -1,17 +1,17 @@
 package logica;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Date;
-import java.sql.Time;
 
 import datatypes.DtActividadDeportiva;
 
 public class ActividadDeportiva {
 	private String nombre;
 	private String descripcion;
-	private int duracion;
-	private float costo;
+	private Float duracion;
+	private BigDecimal costo;
 	private Date fechaReg;
 	private List<Clase> clases = new ArrayList<>();
 	
@@ -19,7 +19,7 @@ public class ActividadDeportiva {
 		super();
 	}
 
-	public ActividadDeportiva(String nombre, String descripcion, int duracion, float costo, Date fechaReg) {
+	public ActividadDeportiva(String nombre, String descripcion, Float duracion, BigDecimal costo, Date fechaReg) {
 		super();
 		this.nombre = nombre;
 		this.descripcion = descripcion;
@@ -44,19 +44,19 @@ public class ActividadDeportiva {
 		this.descripcion = descripcion;
 	}
 
-	public int getDuracion() {
+	public Float getDuracion() {
 		return duracion;
 	}
 
-	public void setDuracion(int duracion) {
+	public void setDuracion(Float duracion) {
 		this.duracion = duracion;
 	}
 
-	public float getCosto() {
+	public BigDecimal getCosto() {
 		return costo;
 	}
 
-	public void setCosto(float costo) {
+	public void setCosto(BigDecimal costo) {
 		this.costo = costo;
 	}
